droid-bind-mounts: Add bind-mount for touch HAL init rc
This will allow us to enable/disable sleep mode display gestures when
the user is in the input group (i.e. nemo/defaultuser under SFOS).

With this I can also stop adding said user to the droid system group
since that was the only thing it was needed for iirc.